refactor: 모든 import 경로를 Path Alias로 마이그레이션
상대 경로 → Path Alias 변환:
- ../../../../utils/* → @utils/*
- ../../../../constants/* → @constants/*
- ../../../../components/* → @components/*
- ../../../../assets/* → @assets/*
- ../../services/* → @services/*
- ../../types/* → @/[page]/types/*
- ../components/* → @/[page]/components/*

영향받은 페이지:
- /works (components, utils, constants, types)
- /people (components, utils)
- /program (components, constants, types)
- /about (components, constants, types)
- /partners (components, constants, types)
- sitemap.ts, page.tsx
