+++ apps/web/app/about/constants/eventHeader.ts
@@ -1,4 +1,4 @@
-import type { EventTitle, RopeFrameConfig } from "../types/eventHeader";
+import type { EventTitle, RopeFrameConfig } from "@/about/types/eventHeader";
 
 export const EVENT_TITLE: EventTitle = {
   mainTitle: "SNU DESIGN WEEK 2025",
