+++ apps/web/app/page.tsx
@@ -1,5 +1,5 @@
 import { Footer } from "@snud2025/ui";
-import PhysicsScene from "../components/PhysicsScene";
+import PhysicsScene from "@components/PhysicsScene";
 
 export default function Home() {
   return (
