+++ apps/web/app/partners/constants/partners.ts
@@ -1,4 +1,4 @@
-import type { Partner } from "../types/partner";
+import type { Partner } from "@/partners/types/partner";
 
 export const PARTNERS: Partner[] = [
   {
