+++ apps/web/app/people/components/PhysicsCell/PhysicsCell.style.ts
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 import { mq } from "@snud2025/ui";
-import { PEOPLE_RESPONSIVE_SCALES } from "../../../../constants/peopleGraphic";
+import { PEOPLE_RESPONSIVE_SCALES } from "@constants/peopleGraphic";
 
 export const PhysicsContainer = styled.div<{
   cellSize: number;
