+++ apps/web/app/works/components/MobileFilter/MobileFilter.tsx
@@ -6,10 +6,10 @@ import {
   getCategoryByIndex,
   getIndexByCategory,
   getCategoryGraphic,
-} from "../../../../utils/categories";
-import { CATEGORIES } from "../../../../constants/categories";
-import type { Category } from "../../../../types/categories";
-import Arrow from "../../../../assets/icons/arrow_down.svg";
+} from "@utils/categories";
+import { CATEGORIES } from "@constants/categories";
+import type { Category } from "@app-types/categories";
+import Arrow from "@assets/icons/arrow_down.svg";
 import { Title } from "@snud2025/ui";
 import * as S from "./MobileFilter.style";
