+++ apps/web/app/works/components/WorksDetailHeader/WorksDetailHeader.tsx
@@ -1,10 +1,10 @@
 import { Body, Subtitle, Title } from "@snud2025/ui";
 import Link from "next/link";
-import MailIcon from "../../../../assets/icons/mail.svg";
-import InstagramIcon from "../../../../assets/icons/instagram.svg";
+import MailIcon from "@assets/icons/mail.svg";
+import InstagramIcon from "@assets/icons/instagram.svg";
 import * as S from "./WorksDetailHeader.style";
-import type { ProjectDetail } from "../../types/projects";
-import { getInstagramUrl } from "../../../../utils/getInstagramUrl";
+import type { ProjectDetail } from "@/works/types/projects";
+import { getInstagramUrl } from "@utils/getInstagramUrl";
 
 interface WorksDetailHeaderProps {
   project: ProjectDetail;
